0.0.2.30 - Tome Position Correction on Delete

+++ Application/AppTrainer/Lores/Delete.cs
@@ -3,7 +3,9 @@ using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using Application.Errors;
+using Application.Interfaces;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 using Persistence;
 
 namespace Application.AppTrainer.Lores
@@ -18,17 +20,40 @@ namespace Application.AppTrainer.Lores
         public class Handler : IRequestHandler<Command>
         {
             private readonly DataContext _context;
-            public Handler(DataContext context)
+            private readonly IUserAccessor _userAccessor;
+            public Handler(DataContext context, IUserAccessor userAccessor)
             {
+                _userAccessor = userAccessor;
                 _context = context;
             }
 
             public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
             {
+                var user = await _context.Users.SingleOrDefaultAsync(x => x.UserName == _userAccessor.GetCurrentUsername());
+
                 var tome = await _context.Tomes.FindAsync(request.Id);
 
+                //  SECURITY
                 if (tome == null)
                     throw new RESTException(HttpStatusCode.NotFound, new { tome = "Not Found" });
+
+                if (tome.AppUserId != user.Id)
+                    throw new RESTException(HttpStatusCode.Forbidden, new { tome = "Forbidden" });
+
+                if (tome.TotalEtudes > 0)
+                    throw new RESTException(HttpStatusCode.Forbidden, new { tome = "Is Not Empty" });
+
+                //  POSITION CORRECTION
+                var startPosition = tome.Position;
+                var tomesQuantity = user.Tomes.Count;
+
+                while (startPosition < tomesQuantity)
+                {
+                    startPosition++;
+                    var nextTome = await _context.Tomes.FirstOrDefaultAsync(x => x.Position == startPosition);
+                    nextTome.Position--;
+
+                }
 
                 _context.Tomes.Remove(tome);
